secretary/requirements.txt
discord.py==0.16.12
dateparser==0.7.0
psycopg2-binary==2.7.6.1